moved utils to astro_utils; changed length of comment lines
