+++ bolides/bdf.py
@@ -20,6 +20,7 @@ from . import MPLSTYLE, ROOT_PATH
 from .utils import reconcile_input
 from .sources import glm_website, usg, pipeline, gmn, csv, remote
 from . import fov_utils as fu
+from .astro_utils import haversine, _distance_metric
 
 _FIRST_COLS = ['datetime', 'longitude', 'latitude', 'source', 'detectedBy',
                'confidenceRating', 'confidence', 'lightcurveStructure', 'energy',
@@ -29,87 +30,6 @@ _FIRST_COLS = ['datetime', 'longitude', 'latitude', 'source', 'detectedBy',
                'impact-e', 'alt', 'vel']
 
 utc = timezone('UTC')
-def haversine(lat1, lon1, lat2, lon2):
-        """
-        Calculates the haversine distance between two points in km.
-        Inputs are in degrees.
-
-        Parameters
-        ----------
-        lat1, lon1 : float
-            Latitude and longitude of the first point in degrees.
-        lat2, lon2 : float
-            Latitude and longitude of the second point in degrees.
-
-        Returns
-        -------
-        float
-            The haversine distance between the two points in kilometers.
-        """
-        R = 6371  # Earth radius in km
-
-        # Convert degrees to radians
-        lat1, lon1, lat2, lon2 = map(np.radians, [lat1, lon1, lat2, lon2])
-
-        dlat = lat2 - lat1
-        dlon = lon2 - lon1
-
-        a = np.sin(dlat / 2)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon / 2)**2
-        c = 2 * np.arcsin(np.sqrt(a))
-        return R * c
-def _distance_metric(x, y, fov_goes, total_time, n):
-    """
-    Measurement of how improbable it is for two bolides to be a distance apart in time and space
-
-    Parameters
-    ----------
-    x, y: tuple
-        Each tuple contains (time_seconds, latitude, longitude). They are rows from the BolideDataFrame.
-        time_seconds is the time in seconds since the start of the dataset.
-        latitude and longitude are in degrees.
-    fov_goes: float
-        The area of the GOES field of view in km^2.
-    total_time: float
-        The total time in seconds over which the bolides were observed.
-    n: int
-        The number of bolides in the dataset.
-
-    Returns
-    -------
-    float
-        The computed distance between the two points, considering both spatial and temporal dimensions. 
-        This estimates the probability that two bolide events would randomly be found as close together in space and time as the two being compared, under a uniform random distribution.
-
-        np.pi*spatial_dist**2
-            This is the area of a circle with radius equal to the spatial distance between the two points (in km²).
-
-        / fov_goes
-            Divides by the total field-of-view area (in km²) of the GOES sensor.
-            This gives the probability that two random points would be within spatial_dist of each other, assuming uniform distribution over the field of view.
-
-        (2*dt/total_time)
-            dt is the absolute time difference between the two events (in seconds).
-            total_time is the total time span of the dataset (in seconds).
-            This gives the probability that two random events would be within dt of each other in time.
-
-        comb(n, 2, exact=False)
-            This is the number of unique pairs you can form from n events.
-            It scales the probability to account for all possible pairs in the dataset.
-
-        This gives the expected number of pairs (out of all possible pairs) that would be at least as close in space and time as the two points being compared, under a random (uniform) distribution.
-
-        Lower values mean the pair is unusually close in space and time (less likely by chance).
-        Higher values mean the pair is not unusually close (more likely by chance).
-
-        Used as the DistanceMetric64 "distance" between the two points for the function get_closest().
-    """
-    
-    t1, x1, y1 = x
-    t2, x2, y2 = y
-    spatial_dist = haversine(x1, y1, x2, y2) 
-    dt = abs(t1 - t2)
-
-    return (np.pi*spatial_dist**2/fov_goes)*(2*dt/total_time) * comb(n, 2, exact=False)
 
 class BolideDataFrame(GeoDataFrame):
     """
@@ -145,7 +65,6 @@ class BolideDataFrame(GeoDataFrame):
         Whether or not to rearrange the columns, if coming from a CSV or Pickle.
     """
     
-
     def __init__(self, *args, **kwargs):
 
         if len(args)==0 and len(kwargs)==0:
@@ -244,7 +163,6 @@ class BolideDataFrame(GeoDataFrame):
         descriptions = pd.read_csv(ROOT_PATH+'/metadata/columns.csv', index_col='column')
         self.descriptions = descriptions[descriptions.sources.isin([source, 'all'])]
         
-
     def annotate(self):
         """Add metadata to bolide detections"""
         from .astro_utils import get_phase, get_solarhour, get_sun_alt
@@ -393,9 +311,20 @@ class BolideDataFrame(GeoDataFrame):
 
     def get_closest(self, datestr=None, lon=None, lat=None, k=1, time_weight=1, loc_weight=1, time_scale=86400, loc_scale=100):
         """
-        Get the n bolides closest to a given iso-format date string and to a given longitude and latitude in degrees. This is calculated using BallTree and a custom distance metric that finds the expected number of pairs of bolides that would be at least as close in space and time as the two points being compared, under a random (uniform) distribution. Effectively, given a time and location, this will find and cluster bolide events that are nearest to it.
+        Get the n bolides closest to a given iso-format date string and to a given
+        longitude and latitude in degrees. This is calculated using BallTree and a
+        custom distance metric that finds the expected number of pairs of bolides
+        that would be at least as close in space and time as the two points being
+        compared, under a random (uniform) distribution. Effectively, given a time
+        and location, this will find and cluster bolide events that are nearest to it.
 
-        If you are more confident about the position or time of the bolide, you can specify weights where higher weights mean you are more confident in that dimension. This will calculate the distance metric as a weighted sum of the time and spatial distances. You can also set the normalization factors for the time and spatial distances, which are used to scale the time and spatial distances to a common scale. The default is 1 day (86400 seconds) for time and 100 km for spatial distance.
+        If you are more confident about the position or time of the bolide, you can
+        specify weights where higher weights mean you are more confident in that
+        dimension. This will calculate the distance metric as a weighted sum of the
+        time and spatial distances. You can also set the normalization factors for
+        the time and spatial distances, which are used to scale the time and spatial
+        distances to a common scale. The default is 1 day (86400 seconds) for time
+        and 100 km for spatial distance.
 
         Parameters
         ----------
@@ -407,9 +336,13 @@ class BolideDataFrame(GeoDataFrame):
         k : int
             Number of closest detections to return.
         time_weight : float
-            Weight to apply to the time difference in the distance metric. Higher values mean you're more confident in the time since you penalize points more heavily if they're farther in time. Default is 1.
+            Weight to apply to the time difference in the distance metric.
+            Higher values mean you're more confident in the time since you penalize
+            points more heavily if they're farther in time. Default is 1.
         loc_weight : float
-            Weight to apply to the spatial distance in the distance metric. Higher values mean you're more confident in the location since you penalize points more heavily if they're farther in distance. Default is 1.
+            Weight to apply to the spatial distance in the distance metric.
+            Higher values mean you're more confident in the location since you penalize
+            points more heavily if they're farther in distance. Default is 1.
         time_scale : float
             Normalization for time in seconds (default 1 day = 86400). 
         loc_scale : float
@@ -426,7 +359,7 @@ class BolideDataFrame(GeoDataFrame):
         elif lon is None and lat is None and datestr is not None:
             return self.get_closest_by_time(datestr, n=k)
         elif lon is None or lat is None:
-            raise ValueError("Must specify both longitude and latitude, or a date string, or all .")
+            raise ValueError("Must specify both longitude and latitude, or a date string, or all three.")
